Initial commit of VFD configuration language parser
For now at least, the configuration language is implemented in the
new H5CL package in src with test code in test/vfd_cl.c.  Updated
the Makefile.am files in src and test for the new files.  All other
file mods / deletions are the result of running autogen.sh and
make distclean before commiting.  Note that this means that you
must configure before you make.

While the current implementation of the VFD configuration language
passes basic smoke checks, the following items remain on the todo
list:

1) Add exhaustive tests for correct parsing of valid configuraiton
   strings.

2) Update the parser to generate useful HDF5 error stacks when a
   syntax error is detected and write test code for same.

3) Integrate the VFD configuration language into the existing
   VFD code to allow string based configuration of VFD stacks.
   Exact details of how this will be done is still TBD, but will
   involve porting the page buffer and encryption VFDs to 1.14.6
   and modifying them to use the VFD configuration language.

Also note that the test code in test/vfd_cl.c is not organized
the way we probably want it, as it seems that Jordan's test
framework mods haven't made it to 1.14.6.